requirements.txt
fitz
pillow
zipfile